+++ grid2op/Backend/EducPandaPowerBackend.py
@@ -62,7 +62,9 @@ class EducPandaPowerBackend(Backend):
 
     """
 
-    def __init__(self, detailed_infos_for_cascading_failures=False):
+    def __init__(self,
+                 detailed_infos_for_cascading_failures=False,
+                 can_be_copied=True):
         """
         Nothing much to do here except initializing what you would need (a tensorflow session, link to some
         external dependencies etc.)
@@ -77,6 +79,7 @@ class EducPandaPowerBackend(Backend):
         Backend.__init__(
             self,
             detailed_infos_for_cascading_failures=detailed_infos_for_cascading_failures,
+            can_be_copied=can_be_copied,
         )
         warnings.warn(
             "This backend is used for demonstration purpose only, you should not use it under any "
@@ -398,14 +401,6 @@ class EducPandaPowerBackend(Backend):
                 1 if bus_id == self.load_to_subid[i] else 2
             )
             i += 1
-
-        # # do not forget storage units !
-        # i = 0
-        # for bus_id in self._grid.storage["bus"].values:
-        #     res[self.storage_pos_topo_vect[i]] = (
-        #         1 if bus_id == self.storage_to_subid[i] else 2
-        #     )
-        #     i += 1
         return res
 
     def generators_info(self):